+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Self-driving car testing".
 - Tests for "No more pythons, please!".
 - Tests for "TicTacToe".
+- Tests for "Number derivation".
 
 ## [3.8.0] - 2022-04-30
 ### Added